Refactored code to enable reuse the basic server communication and encrzption
